fix: dto 기능 분리
